separate the container build from the component build; add support for directories in the containerbuilder
